+++ runtimes/server-interface/chat.ts
@@ -29,6 +29,10 @@ import {
     ListConversationsResult,
     ConversationClickParams,
     ConversationClickResult,
+    GetSerializedChatResult,
+    GetSerializedChatParams,
+    TabBarActionParams,
+    TabBarActionResult,
 } from '../protocol'
 
 /**
@@ -45,6 +49,8 @@ export type Chat = {
     openTab: (params: OpenTabParams) => Promise<OpenTabResult>
     onListConversations: (handler: RequestHandler<ListConversationsParams, ListConversationsResult, void>) => void
     onConversationClick: (handler: RequestHandler<ConversationClickParams, ConversationClickResult, void>) => void
+    onTabBarAction: (handler: RequestHandler<TabBarActionParams, TabBarActionResult, void>) => void
+    getSerializedChat: (params: GetSerializedChatParams) => Promise<GetSerializedChatResult>
     // Notifications
     onSendFeedback: (handler: NotificationHandler<FeedbackParams>) => void
     onReady: (handler: NotificationHandler<void>) => void